WIP: Splitting out chart to use in PDF
