+++ assets/en/report.html
@@ -290,5 +290,8 @@
     <p class="footer">An eviction rate is the number of evictions over the number of renter-occupied households.</div>
 </div>
 {{/if}}
+<div class="page charts">
+    <div class="chart" id="line-chart"></div>
+</div>
 </body>
 </html>
