+++ src/package.json
@@ -11,7 +11,10 @@
     "@types/handlebars": "^4.0.36",
     "aws-sdk": "^2.122.0",
     "axios": "^0.17.1",
+    "canvas-aws-prebuilt": "^2.0.2",
     "chromeless": "^1.3.0",
+    "d3-scale": "^2.0.0",
+    "d3-shape": "^1.2.0",
     "handlebars": "^4.0.10"
   },
   "devDependencies": {
